(PC-29546)[API] refactor: move matching_acceslibre command to script
